requirement.txt
python-dotenv==1.0.1
langchain==0.3.19
langchain-openai==0.3.6